Added blockade and EB cards as ways to grow a city

+++ TBest_Mods/City_Features/Server_AdvanceTurn.lua
@@ -9,16 +9,15 @@ function Server_AdvanceTurn_Start (game, addNewOrder)
 		local structure = {}
 		Cities = WL.StructureType.City
 		for _, territory in pairs(standing.Territories) do
-			if (not territory.IsNeutral) then
-				if not(territory.Structures == nil) then
-					terrMod = WL.TerritoryModification.Create(territory.ID);		
-					structure[Cities] = territory.Structures[WL.StructureType.City] +1;
-					--A cap on how big cities can get
-					if (structure[Cities] < 8 ) then
-						terrMod.SetStructuresOpt   = structure
-						NewOrders[CurrentIndex]=terrMod;
-						CurrentIndex=CurrentIndex+1;
-					end
+			--Can be 0, if a territory has been bombed. We don't want that city to grow.
+			if not(territory.Structures == nil or territory.Structures[WL.StructureType.City] == 0) then
+				terrMod = WL.TerritoryModification.Create(territory.ID);		
+				structure[Cities] = territory.Structures[WL.StructureType.City] +1;
+				--A cap on how big cities can get
+				if (structure[Cities] < 8 ) then
+					terrMod.SetStructuresOpt   = structure
+					NewOrders[CurrentIndex]=terrMod;
+					CurrentIndex=CurrentIndex+1;
 				end
 			end
 		end					
@@ -32,46 +31,66 @@ function Server_AdvanceTurn_Order(game, order, result, skipThisOrder, addNewOrde
 	if (order.proxyType == 'GameOrderAttackTransfer')  then
 		if (result.IsAttack) then
 			if not (game.ServerGame.LatestTurnStanding.Territories[order.To].Structures == nil) then	
-				DefBonus = game.ServerGame.LatestTurnStanding.Territories[order.To].Structures[WL.StructureType.City] * 0.10;
-				attackersKilled = result.AttackingArmiesKilled.NumArmies +  result.AttackingArmiesKilled.NumArmies * DefBonus
-				
-				--Minimum kill 1 attacking army
-				if(attackersKilled == 0) then
-					attackersKilled = 1					
-					--Max armies lost is equal to actualArmies
-					elseif (result.ActualArmies.NumArmies - attackersKilled < 0) then
-					attackersKilled = result.ActualArmies.NumArmies;
-					--Note : At the moment we don't dmg special units
-					--That would be a rare edge case
-					else
-					--round up, always
-					attackersKilled = math.ceil(attackersKilled);
+				if (game.ServerGame.LatestTurnStanding.Territories[order.To].Structures[WL.StructureType.City] > 0) then
+					DefBonus = game.ServerGame.LatestTurnStanding.Territories[order.To].Structures[WL.StructureType.City] * 0.50;
+					attackersKilled = result.AttackingArmiesKilled.NumArmies +  result.AttackingArmiesKilled.NumArmies * DefBonus
+					
+					--Minimum kill 1 attacking army
+					if(attackersKilled == 0) then
+						attackersKilled = 1					
+						--Max armies lost is equal to actualArmies
+						elseif (result.ActualArmies.NumArmies - attackersKilled < 0) then
+						attackersKilled = result.ActualArmies.NumArmies;
+						--Note : At the moment we don't dmg special units
+						--That would be a rare edge case
+						else
+						--round up, always
+						attackersKilled = math.ceil(attackersKilled);
+					end
+					
+					--Write to GameOrderResult	 (result)
+					local NewAttackingArmiesKilled = WL.Armies.Create(attackersKilled) 
+					result.AttackingArmiesKilled = NewAttackingArmiesKilled
+					msg = "The city has " .. tostring(DefBonus*100) .. "% fortification bonus";
+					addNewOrder(WL.GameOrderEvent.Create(order.PlayerID,msg,{}));
 				end
-				
-				--Write to GameOrderResult	 (result)
-				local NewAttackingArmiesKilled = WL.Armies.Create(attackersKilled) 
-				result.AttackingArmiesKilled = NewAttackingArmiesKilled
-				msg = "The city has " .. tostring(DefBonus*100) .. "% fortification bonus";
-				addNewOrder(WL.GameOrderEvent.Create(order.PlayerID,msg,{}));
 			end
 		end
-		
-		--For now, bomb cards reduces cities to 1.
-		elseif(order.proxyType == 'GameOrderPlayCardBomb') then
+	end	
+	--For now, bomb cards reduces cities to 1.
+	if(order.proxyType == 'GameOrderPlayCardBomb') then
 		if not(game.ServerGame.LatestTurnStanding.Territories[order.TargetTerritoryID].Structures == nil) then
 			local structure = {}
 			NewOrders={};
 			Cities = WL.StructureType.City
-			structure[Cities] = game.ServerGame.LatestTurnStanding.Territories[order.TargetTerritoryID].Structures[WL.StructureType.City] -1;
+			structure[Cities] = game.ServerGame.LatestTurnStanding.Territories[order.TargetTerritoryID].Structures[WL.StructureType.City] -2;
 			msg = "City was bombed";
-			if structure[Cities] < 1 then
-				structure = nil;
-				msg = "City was destroyed!"
+			if (structure[Cities] < 1) then
+				structure[Cities] = 0;
+				msg = "The City is destroyed!"
 			end
 			terrMod = WL.TerritoryModification.Create(order.TargetTerritoryID);	
 			terrMod.SetStructuresOpt   = structure
 			NewOrders[1]=terrMod;
 			addNewOrder(WL.GameOrderEvent.Create(order.PlayerID,msg,{},NewOrders));
 		end
+	end	
+	
+	--If we blockade or emergency blockade on a city we own. We build on that city
+	--Maybe, we want this to be a way to make new cities in the future
+	if(order.proxyType == 'GameOrderPlayCardBlockade' or order.proxyType == 'GameOrderPlayCardAbandon') then
+		if not(game.ServerGame.LatestTurnStanding.Territories[order.TargetTerritoryID].Structures == nil) then
+			local structure = {}
+			NewOrders={};
+			Cities = WL.StructureType.City
+			structure[Cities] = game.ServerGame.LatestTurnStanding.Territories[order.TargetTerritoryID].Structures[WL.StructureType.City] +1;
+			msg = "The City have been improved!";
+			terrMod = WL.TerritoryModification.Create(order.TargetTerritoryID);	
+			terrMod.SetStructuresOpt   = structure
+			NewOrders[1]=terrMod;
+			addNewOrder(WL.GameOrderEvent.Create(order.PlayerID,msg,{},NewOrders));
+			skipThisOrder(WL.ModOrderControl.SkipAndSupressSkippedMessage);
+		end
 	end
 end
+
